refactor: extract ACP close session flow
